Tests/LibWeb/Text/input/input-value-does-not-force-layout.html

<!DOCTYPE html>
<input id="input" />
<div id="other"></div>
<script src="include.js"></script>
<script>
    test(() => {
        input.focus();
        input.offsetWidth;

        other.style.padding = "10px";
        input.value = 42;
        println(`layout is ${internals.layoutIsUpToDate ? "" : "not "}up to date`);
    });
</script>
