Tests/LibWeb/Text/expected/input-value-does-not-force-layout.txt

layout is not up to date
